requirements.txt
langchain_community
langchain_openai
openai
langgraph
streamlit